test: unblock main ci checks
