+++ src/gateway/plugin-node-capability-types.ts
@@ -1,0 +1,11 @@
+export type PluginNodeCapabilitySurface = {
+  surface: string;
+  ttlMs?: number;
+  scopeKey?: string;
+};
+
+export type PluginNodeCapabilityClient = {
+  pluginSurfaceUrls?: Record<string, string>;
+  pluginNodeCapabilitySurfaces?: Record<string, PluginNodeCapabilitySurface>;
+  pluginNodeCapabilities?: Record<string, { capability: string; expiresAtMs: number }>;
+};
